Revert "Update index.yaml" (#499)
This reverts commit 0d24f68cf8440a84e0ac677fc6e1eb05c4a32c37.

MCK release 1.3.0 was broken. This commit tries to remove that version from the helm index so that we can re-release that version.
Our releaser scripts checks if helm search repo returns the version and if it does, the releaser script doesn't upload the chart tar.
This PR would remove the 1.3.0 version from helm index so that it can again be created.